+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Resolve depth of mates from transposition table
 - Use SIMD operations for pawn heuristics
 
+### Removed
+
+- Allocator removed in move generation and search
+
 ## [4.2] - 2026-01-17
 
 ### Added